refactor axis utils
closes #455
